gaussian_random_kernel and mkldnn-onednn renaming
